requirements.txt
requests~=2.27.1
python-dotenv~=0.19.2
google-api-python-client~=2.36.0
google-auth-httplib2~=0.1.0
google-auth-oauthlib~=0.4.6
